+++ extensions/dashlane-vault/src/helper/exec.ts
@@ -1,0 +1,4 @@
+import { exec } from "child_process";
+import util from "util";
+
+export const execPromise = util.promisify(exec);