+++ extensions/github/src/helpers/errors.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function getErrorMessage(error: any) {
   if (error?.response?.errors) {
     return error.response.errors[0].message;