+++ extensions/reclaim-ai/src/notifications.tsx
@@ -14,11 +14,10 @@ import { useMemo } from "react";
 import { useEvent } from "./hooks/useEvent";
 import { ApiResponseEvents, ApiResponseMoment } from "./hooks/useEvent.types";
 import { useUser } from "./hooks/useUser";
-import { CalendarAccount } from "./types/account";
 import { Event } from "./types/event";
 import { NativePreferences } from "./types/preferences";
 import { miniDuration } from "./utils/dates";
-import { eventColors, truncateEventSize } from "./utils/events";
+import { eventColors, filterMultipleOutDuplicateEvents, truncateEventSize } from "./utils/events";
 import { parseEmojiField } from "./utils/string";
 
 type EventSection = { section: string; sectionTitle: string; events: Event[] };
@@ -77,28 +76,20 @@ export default function Command() {
     Accept: "application/json",
   };
 
-  const { data: accountData, isLoading: isLoadingAccounts } = useFetch<CalendarAccount[]>(`${apiUrl}/accounts`, {
-    headers: fetchHeaders,
-    keepPreviousData: true,
-  });
-
   const { data: eventsResponse, isLoading: isLoadingEvents } = useFetch<ApiResponseEvents>(
     `${apiUrl}/events?${new URLSearchParams({
       sourceDetails: "true",
       start: format(startOfDay(new Date()), "yyyy-MM-dd"),
       end: format(addDays(new Date(), 2), "yyyy-MM-dd"),
-      ...(accountData && {
-        calendarIds: accountData.flatMap(({ connectedCalendars }) => connectedCalendars.map(({ id }) => id)).join(","),
-      }),
+      allConnected: "true",
     }).toString()}`,
     {
-      execute: !!accountData && !isLoadingAccounts,
       headers: fetchHeaders,
       keepPreviousData: true,
     }
   );
 
-  const eventData = eventsResponse;
+  const eventData = filterMultipleOutDuplicateEvents(eventsResponse);
 
   const showDeclinedEvents = useMemo(() => {
     return !!currentUser?.settings.showDeclinedEvents;