+++ extensions/apple-reminders/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Apple Reminders Changelog
 
+## [Bug fixes] - 2023-11-13
+
+- Add notifications to reminders when they're set for a specific time
+- Don't reset the due date, priority, and list fields when creating a new reminder
+
+## [Today and Scheduled views] - 2023-11-03
+
+- Add support for today and scheduled views, similar to those in the native Reminders app.
+
 ## [Recurring remidners and better group by dates] - 2023-10-24
 
 - Add support for basic recurring reminders in `Create Reminder` and `Quick Add Reminder`