+++ extensions/reclaim-ai/CHANGELOG.md
@@ -1,5 +1,23 @@
 # reclaim Changelog
 
+## [Fix] - 2023-11-13
+
+- Create task is now using the default task visibility rather than always private
+
+## [Update] - 2023-11-07
+
+- Introduce new functionality to filter out synchronized events managed by Reclaim that are part of multiple calendars.
+
+## [Update] - 2023-11-02
+
+- Changed task filter to 'Archived' only
+- Changed order of task status
+
+## [Update] - 2023-10-31
+
+- Refactor of Event Fetching Process
+- Bug Fix in the Menu Bar Component
+
 ## [Update] - 2023-10-24
 
 - Revert filter logic